requirements.txt
django-colorful
django-filter
django-markdown-deux
django-tagulous
Pillow
pypandoc
tablib[pandas]
ujson
